move priority queue to new common crate (#2304)
