+++ lib/collection/Cargo.toml
@@ -46,6 +46,7 @@ url = { version = "2", features = ["serde"] }
 validator = { version = "0.16", features = ["derive"] }
 actix-web-validator = "5.0.1"
 
+common = {path = "../common"}
 segment = {path = "../segment"}
 api = {path = "../api"}
