+++ lib/common/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "common"
+version = "0.1.0"
+authors = [
+    "Andrey Vasnetsov <vasnetsov93@gmail.com>",
+    "Qdrant Team <info@qdrant.tech>",
+]
+license = "Apache-2.0"
+edition = "2021"
+
+[dependencies]
+serde = "~1.0"
