+++ lib/common/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod fixed_length_priority_queue;
