+++ lib/segment/src/index/hnsw_index/graph_layers.rs
@@ -1,6 +1,7 @@
 use std::cmp::max;
 use std::path::{Path, PathBuf};
 
+use common::fixed_length_priority_queue::FixedLengthPriorityQueue;
 use itertools::Itertools;
 use serde::{Deserialize, Serialize};
 
@@ -14,7 +15,6 @@ use crate::index::hnsw_index::graph_links::GraphLinksConverter;
 use crate::index::hnsw_index::point_scorer::FilteredScorer;
 use crate::index::hnsw_index::search_context::SearchContext;
 use crate::index::visited_pool::{VisitedList, VisitedPool};
-use crate::spaces::tools::FixedLengthPriorityQueue;
 use crate::types::PointOffsetType;
 use crate::vector_storage::ScoredPointOffset;
