+++ lib/segment/src/index/hnsw_index/search_context.rs
@@ -1,9 +1,9 @@
 use std::collections::BinaryHeap;
 use std::iter::FromIterator;
 
+use common::fixed_length_priority_queue::FixedLengthPriorityQueue;
 use num_traits::float::FloatCore;
 
-use crate::spaces::tools::FixedLengthPriorityQueue;
 use crate::types::ScoreType;
 use crate::vector_storage::ScoredPointOffset;
